Refactor units annotator and fix annotation of units to use pre-defined IDs instead of aliases
