+++ src/python/annotate_sbml_units.py
@@ -2,7 +2,7 @@ import sys
 import argparse
 from pathlib import Path
 import traceback
-import unit_annotation_helpers as uah
+import sbml_units_annotator as ua
 
 def main():
     parser = argparse.ArgumentParser(description="Annotate the units of the SBML file with the specifications of the annotations file")
@@ -23,7 +23,8 @@ def main():
             sbml_file = f_in
             annotations_file = f_ann
             out_file = f_out
-            uah.annotateUnits(sbml_file, annotations_file, str(out_file))
+            annotator = ua.sbmlUnitsAnnotator()
+            annotator.annotateUnits(sbml_file, annotations_file, str(out_file))
         else:
             print(f'[{f_out}] already exists, use -f to force conversion')
     except Exception as e:
